+++ src/CodePreprocessor/PreprocessorInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Sstalle\php7cc\CodePreprocessor;
+
+interface PreprocessorInterface
+{
+    /**
+     * @param string $code
+     *
+     * @return string
+     */
+    public function preprocess($code);
+}