+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Version 1.2-dev
+## Version 1.2.0
 * Fixed some false positives related to yield usage in expression context
 * Fixed false positives for nested by-reference foreach loops wrapped in functions or class methods
 * Fixed function redeclaration false positives when a declaration is wrapped in a function_exists condition